move MambaCacheManager and MambaHybridCacheManager to a separate file
